+++ mycli/main.py
@@ -94,7 +94,7 @@ class MyCli(object):
     ]
 
     default_config_file = os.path.join(PACKAGE_ROOT, 'myclirc')
-
+    pwd_config_file = os.path.join(os.getcwd(), ".myclirc")
 
     def __init__(self, sqlexecute=None, prompt=None,
             logfile=None, defaults_suffix=None, defaults_file=None,
@@ -114,7 +114,7 @@ class MyCli(object):
 
         # Load config.
         config_files = ([self.default_config_file] + self.system_config_files +
-                        [myclirc])
+                        [myclirc] + [self.pwd_config_file])
         c = self.config = read_config_files(config_files)
         self.multi_line = c['main'].as_bool('multi_line')
         self.key_bindings = c['main']['key_bindings']
@@ -1093,7 +1093,7 @@ def cli(database, user, host, port, socket, password, dbname,
             database = uri.path[1:]  # ignore the leading fwd slash
         if not user:
             user = unquote(uri.username)
-        if not password:
+        if not password and uri.password is not None:
             password = unquote(uri.password)
         if not host:
             host = uri.hostname